Rearrange some function defs
